correct 'props'

+++ braph2genesis/src/ds/Element.m
@@ -1425,7 +1425,7 @@ classdef Element < Category & Format & matlab.mixin.Copyable
                                 json_str = [json_str '}']; %#ok<AGROW>
                                 struct{i}.props{prop}.value = json_str;
                             case Format.NET
-                                struct{i}.props{prop}.value = net_to_onnx(value);
+                                struct{i}.props{prop}.value = Element.net_to_onnx(value);
                         end
                     end
                     struct{i}.props{prop}.seed = el.getPropSeed(prop);
@@ -1450,7 +1450,7 @@ classdef Element < Category & Format & matlab.mixin.Copyable
             % manages special case when only one element
             if length(struct) == 1
                 struct_tmp{1}.class = struct.class;
-                if isfield(struct, 'prop')
+                if isfield(struct, 'props')
                     struct_tmp{1}.props = struct.props;
                 end
                 struct = struct_tmp;
@@ -1501,7 +1501,7 @@ classdef Element < Category & Format & matlab.mixin.Copyable
                             case Format.CELL
                                 el.props{prop}.value = eval(value);
                             case Format.NET
-                                el.props{prop}.value = onnx_to_net(value);
+                                el.props{prop}.value = Element.onnx_to_net(value);
                         end
                     end
                     el.props{prop}.seed = uint32(struct{i}.props(prop).seed);
@@ -1527,19 +1527,23 @@ classdef Element < Category & Format & matlab.mixin.Copyable
             w = warning('query','MATLAB:mir_warning_unrecognized_pragma');
             warning('off', 'MATLAB:mir_warning_unrecognized_pragma');
 
-            directory = [fileparts(which('test_braph2')) filesep 'trial_nn_from_matlab_to_be_erased'];
-            if ~exist(directory, 'dir')
-                mkdir(directory)
+            if isa(net, 'network') || iscell(net)
+                onnx_str = '';
+            else
+                directory = [fileparts(which('test_braph2')) filesep 'trial_nn_from_matlab_to_be_erased'];
+                if ~exist(directory, 'dir')
+                    mkdir(directory)
+                end
+                filename = [directory filesep 'nn_from_matlab_to_be_erased.onnx'];
+
+                exportONNXNetwork(net, filename);
+
+                fileID = fopen(filename);
+                onnx_str = num2str(fread(fileID));
+                fclose(fileID);
+
+                rmdir(directory, 's')
             end
-            filename = [directory filesep 'nn_from_matlab_to_be_erased.onnx'];
-
-            exportONNXNetwork(net, filename);
-
-            fileID = fopen(filename);
-            onnx_str = {fread(fileID)};
-            fclose(fileID);
-
-            rmdir(directory, 's')
             warning(w.state, 'MATLAB:mir_warning_unrecognized_pragma')
         end
         function net = onnx_to_net(onnx_str, varargin)
@@ -1559,32 +1563,29 @@ classdef Element < Category & Format & matlab.mixin.Copyable
             warning('off', 'MATLAB:mir_warning_unrecognized_pragma');
             w_nnet = warning('query','nnet_cnn:internal:cnn:analyzer:NetworkAnalyzer:NetworkHasWarnings');
             warning('off','nnet_cnn:internal:cnn:analyzer:NetworkAnalyzer:NetworkHasWarnings');
-
-            directory = [fileparts(which('test_braph2')) filesep 'trial_nn_from_braph_to_be_erased'];
-            if ~exist(directory, 'dir')
-                mkdir(directory)
-            end
-            filename = [directory filesep 'nn_from_braph_to_be_erased.onnx'];
-
-            fileID = fopen(filename, 'w');
-            fwrite(fileID, onnx_str{1});
-            fclose(fileID);
-
-            if length(varargin) == 3
-                format = varargin{1};
-                type = varargin{2};
-                class_name = varargin{3};
-                net = importONNXNetwork(filename, 'OutputLayerType', type, 'Classes', class_name);
-            elseif length(varargin) == 2
-                format = varargin{1};
-                type = varargin{2};
-                net = importONNXNetwork(filename, 'OutputLayerType', type);
+            
+            if isempty(onnx_str)
+                if BRAPH2.installed('NN', 'msgbox')
+                    net = network();
+                else
+                    net = {};
+                end
             else
-                lgraph = importONNXLayers(filename);
-                net = assembleNetwork(lgraph);
-            end
+                directory = [fileparts(which('test_braph2')) filesep 'trial_nn_from_braph_to_be_erased'];
+                if ~exist(directory, 'dir')
+                    mkdir(directory)
+                end
+                filename = [directory filesep 'nn_from_braph_to_be_erased.onnx'];
 
-            rmdir(directory, 's');
+                fileID = fopen(filename, 'w');
+                fwrite(fileID, str2double(onnx_str));
+                fclose(fileID);
+
+                lgraph = importONNXLayers(filename);
+                
+                net = assembleNetwork(lgraph);
+                rmdir(directory, 's');
+            end
             warning(w_matlab.state, 'MATLAB:mir_warning_unrecognized_pragma');
             warning(w_nnet.state,'nnet_cnn:internal:cnn:analyzer:NetworkAnalyzer:NetworkHasWarnings');
         end
